pyproject.toml
[tool.poetry]
name = "CS_ViT"
version = "0.1.0"
description = ""
authors = ["Mine268 <mine268@foxmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
torch = {version = "^2.6.0", extras = ["cuda"]}
numpy = "^2.2.3"
einops = "^0.8.1"
tensorboard = "^2.19.0"
matplotlib = "^3.10.0"
typing-extensions = "^4.12.2"
transformers = "^4.49.0"
ipython = "^8.32.0"
tqdm = "^4.67.1"
scipy = "^1.15.2"
PyYAML = "^6.0.2"
debugpy = "^1.8.12"
jupyterlab = "^4.3.5"
typeguard = "^4.4.2"
open3d = "^0.19.0"
smplx = "^0.1.28"
torchvision = "^0.21.0"
pytest = "^8.3.4"
opencv-python = "^4.11.0"
kornia = "^0.8.0"
nvitop = "^1.4.2"
torchsummary = "^1.5.1"
torchsummaryX = "^1.3.0"
pycocotools = "^2.0.8"
setuptools = "^75.8.2"
backports-lzma = "^0.0.14"
pylzma = "^0.5.0"
pyturbojpeg = "^1.7.7"
plyfile = "^1.1"
ipykernel = "^6.29.5"
ipympl = "^0.9.7"
peft = "^0.14.0"
h5py = "^3.13.0"
memory-profiler = "^0.61.0"
pympler = "^1.1"
memray = "^1.16.0"
colorama = "^0.4.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
